First stage of deprecating EWS XML items
The initial idea with the XML item classes was to create a hierarchy of
classes that could read and write SOAP requests and responses. After
doing some work with it it turned out that another layer was needed to
set and retreive the required data from the XML documents which made
thing overcomplicated.

The new approach is to move the responsibility for creating and reading
the XML request/response into the request classes. For commonly used
things like the folder identifier or shape wrapper classes were created
that also have serialization methods.

The first stage switches the request XML generation to the new
method. Response processing will follow.