Rename EwsFolder to EwsMailFolder
Also add possibility to upcast an EwsFolderBase object that holds mail
folder data into a trueEwsMailFolder. This charade is needed because
when retreiving a folder its type is not yet known. Therefore a generic
folder is returned. However internally the private data is allocated
based on the actual folder type. The upcast allows to safely turn the
generic object into a specialized one based on the underlying data.