Add wrapper class for EWS items
The wrapper class is meant to hold items such as e-mails, contacts,
tasks or calendar items.

Its design is based on the EwsFolderBase class with some
differences. Instead of creating explicit fields for each value returned
by EWS the class uses a generic hash of item fields. The reasoning
behind this is the large variety of fields available for items. Creating
class fields for each of them would be a waste of memory as most of them
are rarely used anyway not to mention that a hash allows better control
over the actual fields that are available due to the item shape used for
the EWS request.

The item class is also not a QObject any more as it is meant to be a
PIMPL class with shared data. QObject-derived classes are not meant to
be copied. This caused removal of the self-update mechanism, which
wasn't really used anyway and is not expected to be used in the future.

Another difference compared to the folder classes is lack of
item-specific derived classes. These were only used to add item-specific
fields, which are now handled by the value hash.

The general idea is to refactor the EwsFolderBase & descendants into a
EwsFolder class derived from EwsItemBase once the design of EwsItem
proves it's usefulness.