Use special method to clone PIMPL objects
Since the private data pointer is declared in the base class it is not
aware of the size of the descendant private data classes, which may
contain additional fields. In such case a COW clone of the shared data
is dangerous as it only allocates an object of the base private
class. Any writes to the descendant-only fields will cause heap
corruption.

A resolution has been copied from Akonadi classes which use a virtual
method in the private class to allocate itself. The virtual nature of
the method guarantees that the cloned object will be allocated by the
right class and with the right size.