Pass response class to response item handler
When collecting multiple items Exchange will return separate
XxxResponseMessage elements for each item. This means that an operation
on each item may independently fail. In order to handle such situations
don't store the response class globally but rather pass it to the
response item handler for it to decide what to do with this problem.