Refactor request response processing
Initially the request framework assumed the existence of a single
response item. After implementing some more requests in turned out that
some of them can return multiple responses - one per retrieved item. In
order to implement proper support for this introduce the concept of
response items, which are meant to contain information about the status
of each response and any items it contains.