Add server version support in requests
This comes in two parts. When preparing a request the EwsRequest class
will use the lowes possible version (Exchange 2007) unless a different
version is set.

When reading a response the server version is parsed and stored in the
EwsClieng object. This allows the resource implementation to detect if
some features available only in later Exchange server versions can be used.