Rework namespace support in XML reader
The previous version looked great on paper, but when trying to run it
the program crashes. The reason for the crash is wrong initialization
order between the XML reader items table and the namespace URI
constants. The effect is that the item array gets initialized before the
constants which causes an attempt to initialize the items with a null
string and in turn a crash.

To fix the problem the namespace URI is now provided as an argument to
the parsing method call. This introduces a limitation to a single
namespace within one element. Fortunately in the EWS world this will
always be the case.