Handle partial item move success/failure
When moving multiple items it can happen that moving some items will
succeed while it will fail for others. Such situation can leave Akonadi
in an invalid state as it will assume that the items were moved on the
back-end.

If the move operation fails for some items attempt to partially revert
the operation by moving them back into their original collections.

In the long term a more correct approach would be to issue a special
sync request for the source folder which would check if the item(s) in
question still exist and move them back if yes or delete them otherwise.