Yet another change to handling partial move failure
This attempt will trigger updates to both source and destination
collections in case at least one of the moves fail. Since however the
sync operation is incremental it will normally not pick up the changes,
unless a full sync is executed. In order to avoid doing a potential
expensive full sync the EwsFetchItems job has gained the ability to
explicitly query the existence of given items aside from a regular sync
query.