Add a second resource for sending e-mail
Because of current KMail limitations it is necessary to create a
separate resource for sending e-mail over EWS. Its implementation is
based on a DBus interface published by the main EWS resource, which
performs all EWS operations.

Creation of this resource is configurable and currently enabled by
default. The intention is to introduce the necessary changes in KMail so
that the default can be changed in a future release.