Fix envelope retrieval for some messages
Envelope retrieval relies on retrieving the full set of headers from the
message. For this purpose the request retrieves the
InternetMessageHeaders along with a few other fields for each
message. Unfortunately for some strange reasons the
InternetMessageHeaders item is never returned for some messages.

In order to fix this stop relying on the above field and manually
retrieve all the necessary header fields like the date, message id and
referenced messages.