Add the ability to ignore events to the item fetch job
Similar to the subscription manager the item fetch job also needs to be
able to ignore item operations performed by the resource itself. This is
most important for delete and move operations as without this feature
the feedback loop will cause the incremental request to fail and a full
sync to be forced.