Enable NTLMv2 for EWS requests
By default KIO will only attempt to negotiate NTLMv1 authentication. The
problem is that some newer Exchange servers ban NTLMv1 and it is not
always possible to safely fallback to NTLMv2. Therefore enable support
for NTLMv2 by default. This should be safe as Exchange servers are
usually modern enough to support NTLMv2.