Add workaround for incoming mail filtering
Current Akonadi versions only consider inbox folders for filtering the
incoming mail. A folder is considered an inbox folder when its remote
identifier is "INBOX", which is a correct assumption for IMAP, but not
necessarily for other backends.

In order to work around the problem the EWS resource will pretend that
the remote identifier of the inbox folder is "INBOX", while the actual
identidier is kept internal to the resource.

This workaround can be conditionally compiled-in and will be removed
once the filtering logic in Akonadi is fixed.