Fix inbox folder subscription
Due to the inbox filtering workaround the real folder id for inbox is
stored internally and the string "INBOX" is used at Akonadi level. The
substitution is done in the EwsId class.

During resource startup there is a race condition, whereby the
EwsSubscribedFoldersJob tries to validate the list of subscribed folders
before the initial GetFolder request comes back with the real inbox
folder id. This causes the inbox id in the validation request to be
empty and causes the inbox to be removed from the subscription list.

This change adds a workaround for it in that when the real inbox id is
not known the distinguished id is used.