Work around broken SyncFolderHierarchy request
It seems that at least some Exchange versions will error out with an
Internal Server Error when a SyncFolderHierarchy request is executed on
a mailbox with a large number of folders. The suspition is that this can
be due to significant size of the reply sent back to the client.

The change attempts to work around that by making a second attempt at
the SyncFolderHierarchy request only requesting the folder identifiers
and fetching the rest of the details in subsequent GetFolder request in
batches per 50 folders.