Finalize folder synchronisation implementation
This involves splitting the fill and incremental synchronisation logic
into two separate jobs. This simplifies the implementation of each of
the jobs as the shared code is not as significant as expected.

The incremental sync logic has been redesigned to handle most of the
cases. For any complicated changes the incremental sync will fail
which (similarly to the item sync) will cause a fallback to a full sync.