Add workaround for Qt 5.7 build issue
In Qt 5.7 the qdbuscpp2xml tool's behavior changed and it no longer
looks for include files in the directory of the output
file (QTBUG-55148). This breaks creation of the DBus interface as it
contains conditionally-compiled methods for which the define is
contained in an external header.

As a workaround the ewsresource.h header is now also preprocessed by
CMake in order to put the relevant definition directly into it.