Fix streaming notification on network change
The streaming notification mode keeps a connection with the Exchange
server to receive instant updates to the mailbox. You would expect that
if the connection is broken due to network issues the EWS resource would
be notified of this fact, but the sad reality is that it is not. This
causes the resource to wait forever on a dead connection and no
notifications are ever received.

The change introduces a timer that waits for the regular sends regular
empty 'keep alive' responses that are sent by Exchange. If no message is
received within the timeout the existing connection is terminated and
restarted.