Allow using custom user agent string
In order to enforce some organization policies that disallow use of
non-authorized e-mail clients the Exchange server has the ability to ban
EWS clients other than the explicitly allowed ones. This is done by
analyzing the HTTP User-Agent string and matching it agains a whitelist.

This change adds a user-configurable option to allow emulation of user
agent strings sent to the Exchange server. A list of known e-mail
clients is provided along with the possibility to manually enter a user
agent string.