Distinguish between no domain and empty domain
The default behaviour of NTLM authentication when no domain name is
provided is to use the name returned by the server. In some cases
however it is necessary to force using an empty domain name. The change
adds a check box that enables sending the domain name even if empty.

In order to work correctly an accompanying change in KIO is
needed (commit 20eac1f), which is expected to be released with KF 5.28.

This fixes bug #3.