Workaround for date when creating items in the Sent Items folder
The "Sent Items" folder is a little "special" when it comes to creating
items. Unlike other folders when creating items there the creation
date/time is always set to the current date/time instead of the value
from the MIME Date header. This causes mail that was copied from other
folders to appear with the current date/time instead of the original
one.

To work around this create the item in the "Drafts" folder first and
then move it to "Sent Items".

Fixes bug #5.