Keep mailbox subscription persistent
The mailbox subscription is guaranteed to be valid for 10 minutes of
inactivity. Use that and attempt to reuse an existing subscription
instead of always creating a new one.

This will be most useful for temporary connection losses or resource
restarts as it will allow instant notification about any changes on the
Exchange mailbox within the offline period.