Remove remote tags that are not present locally
This is not the ideal solution as it can cause tags to disappear when
multiple Akonadi instances are sharing access to an Exchange
account. Unfortunately due to broken tag support it is necessary to
avoid situations when a tag is erroneously duplicated on Akonadi side,
propagated onto Exchange and then once Akonadi cleans up the EWS
resource tries to restore that tag from the server causing errors.

This will need to be reviewed once Akonadi tag support is improved.