Retry authentication with last redirected URL
When performing autodiscovery it is common to see a lot of redirects
ending up with a 401 Unauthorized response once the final URL is
hit. When retrying with credentials make sure to use the actual URL that
resulted in the 401 response rather than the initial autodiscovery URL
as doing the latter may trigger a warning from KIO that an attempt was
made to pass credentials to a host that never asked for them.