Fix handling of event subscription errors
When Exchange doesn't like the subscription id it will respond with an
error response. However this response will be a successful HTTP response
hence KJob::error() will be set to 0, but the content will contain the
error. This means that handling the error when KJob::error() is non-zero
is wrong and it should be moved.

This fixes attempts to reuse existing subscriptions.