Add more refined state management for HTTP request
KIO HTTP posts its request in three phases:
 - initial headers
 - Content-Length header + end of headers
 - content

The lack of the Content-Length header in the initial header set caused a
failure as the code was not prepared for a second batch of headers. To
fix this introduce some simple state management so that all headers will
get parsed no matter when they come.