Switch to using XQuery for matching requests
While the regexp-based approach was fine for simple requests, complex
XML requests can be a pain to match with regular expressions as the same
logical content can be expressed in a number of ways.

To improve that switch to using XQuery to perform matching as it allows
to use XPath expressions to perform XML-aware matching.