Fix broken encoding for sent e-mails
Exchange doesn't just store whatever MIME content that was sent to it -
it will parse it and send further the version assembled back from the
parsed parts. It seems that this parsing doesn't work well with the
quoted-printable encoding, which KMail prefers. This results in
malformed encoding, which the sender doesn't even see.

As a workaround force encoding of the body (or in case of multipart -
all parts) to Base64.