Initial sample isolated test for Akonadi EWS
This introduces a framework that is based on existing set-up in other
Akonadi resources and is designed to test Akonadi EWS in a more
realistic environment.

This works by launching the Akonadi EWS resource in an isolated Akonadi
server instance in a fully controlled temporary and volatile
environment. The Akonadi EWS resource communicates with a fake Exchange
EWS server that can be programmed to react to EWS request in a defined
way allowing to playback various scenarios that would otherwise be
difficult to trigger on a real Exchange server.

The initial test only launches Akonadi EWS and sets it online for a
brief moment. Further tests will be developed to verify other real-world
scenarios including error cases.